rename fee-> fee_proof
